refactor: 뷰 책임 분리
