+++ src/main/kotlin/racingcar/RacingGame.kt
@@ -1,24 +1,23 @@
 package racingcar
 
+import racingcar.domain.Car
+import racingcar.domain.Cars
+import racingcar.view.readNames
+import racingcar.view.showCarsPosition
+import racingcar.view.showCountInputMessage
+import racingcar.view.showNameInputMessage
+
 class RacingGame {
     fun play() {
-        println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
-        val input = readln()
-        val names = input.split(",").map { it.trim() }
-        val values = mutableListOf<Car>()
-        for (name in names) {
-            values.add(Car(name))
-        }
-        val cars = Cars(values)
+        showNameInputMessage()
+        val cars = Cars(readNames().map { Car(it) })
 
-        println("시도할 회수는 몇회인가요?")
-        val count: Int = readln().toInt()
+        showCountInputMessage()
+        val count: Int = readln().toIntOrNull()?: throw IllegalArgumentException()
+
         for (i in 1..count) {
             cars.moveAll()
-            for (car in cars.values) {
-                println("${car.name} : " + "-".repeat(car.position))
-            }
-            println()
+            showCarsPosition(cars)
         }
 
         val winners = cars.findWinners()
