+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,0 +1,5 @@
+package racingcar.view
+
+fun readNames(): List<String> {
+    return readln().split(",").map { it.trim() }
+}
