+++ src/main/kotlin/racingcar/view/OutputView.kt
@@ -1,0 +1,18 @@
+package racingcar.view
+
+import racingcar.domain.Cars
+
+fun showNameInputMessage() {
+    println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
+}
+
+fun showCountInputMessage() {
+    println("시도할 회수는 몇회인가요?")
+}
+
+fun showCarsPosition(cars: Cars) {
+    for (car in cars.values) {
+        println("${car.name} : " + "-".repeat(car.position))
+    }
+    println()
+}
