+++ src/test/kotlin/racingcar/CarTest.kt
@@ -3,6 +3,7 @@ package racingcar
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
+import racingcar.domain.Car
 
 class CarTest : AnnotationSpec() {
